Ticket #4953 - allow bigint type for some more fields
git-svn-id: file:///data/svn/enum/server/trunk@15982 b60c293d-5e0d-0410-87b2-fb9704ccd4b5

+++ src/fredlib/requests/model_request_property_value.cc
@@ -6,7 +6,7 @@ DEFINE_BASIC_FIELD(ModelRequestPropertyValue, Database::DateTime, requestTimeBeg
 DEFINE_BASIC_FIELD(ModelRequestPropertyValue, int, requestServiceId, m_requestServiceId, table_name, "request_service_id", .setNotNull())
 DEFINE_BASIC_FIELD(ModelRequestPropertyValue, bool, requestMonitoring, m_requestMonitoring, table_name, "request_monitoring", .setNotNull())
 DEFINE_PRIMARY_KEY(ModelRequestPropertyValue, unsigned long long, id, m_id, table_name, "id", .setDefault())
-DEFINE_BASIC_FIELD(ModelRequestPropertyValue, int, requestId, m_requestId, table_name, "request_id", .setNotNull())
+DEFINE_BASIC_FIELD(ModelRequestPropertyValue, unsigned long long, requestId, m_requestId, table_name, "request_id", .setNotNull())
 DEFINE_BASIC_FIELD(ModelRequestPropertyValue, int, propertyNameId, m_propertyNameId, table_name, "property_name_id", .setNotNull())
 DEFINE_BASIC_FIELD(ModelRequestPropertyValue, std::string, value, m_value, table_name, "value", .setNotNull())
 DEFINE_BASIC_FIELD(ModelRequestPropertyValue, bool, output, m_output, table_name, "output", .setDefault())
